fix(tests): avoid backend dependency imports in root CI
